@@ -1,16 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
         <PackageReference Include="AutoMapper" Version="8.1.1" />
-        <PackageReference Include="Microsoft.AspNetCore.App">
-          <PrivateAssets Condition="'%(PackageReference.Version)' == ''">all</PrivateAssets>
-          <Publish Condition="'%(PackageReference.Version)' == ''">true</Publish>
-        </PackageReference>
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <FrameworkReference Include="Microsoft.AspNetCore.App" />
     </ItemGroup>
 
     <ItemGroup>
