@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -5,8 +6,7 @@ namespace Common.Contracts
 {
     public interface IMessageQueue
     {
-        Task SendMessageAsync<T>(T Messsage);
-        Task SendMessagesAsync<T>(List<T> Messsage);
-        
+        Task SendMessageAsync<T>(T Messsage, string label = null, DateTime? enqueueTime = null);
+        Task SendMessagesAsync<T>(List<T> Messsage, string label = null, DateTime? enqueueTime = null);
     }
 }
